chore: split utils to single files - FUI-1967
