+++ .genx/utils/index.js
@@ -1,0 +1,19 @@
+const makeDirectory = require('./makeDirectory');
+const registerPartials = require('./registerPartials');
+const generateRoute = require('./generateRoute');
+const validateRoute = require('./validateRoute');
+const generateEmptyCsv = require('./generateEmptyCsv');
+const formatRouteData = require('./formatRouteData');
+const parseJSONArgument = require('./parseJSONArgument');
+const excludeFrameworks = require('./excludeFrameworks');
+
+module.exports = {
+  makeDirectory,
+  registerPartials,
+  generateRoute,
+  validateRoute,
+  generateEmptyCsv,
+  formatRouteData,
+  parseJSONArgument,
+  excludeFrameworks,
+};
