+++ .genx/utils/makeDirectory.js
@@ -1,0 +1,9 @@
+const { existsSync, mkdirSync } = require('node:fs');
+
+const makeDirectory = (directory) => {
+  if (!existsSync(directory)) {
+    mkdirSync(directory);
+  }
+};
+
+module.exports = makeDirectory;
